Moved `sha256`, `hash` and `hashMessage` to the `crypto` package
