+++ packages/account/src/hdwallet/hdwallet.ts
@@ -1,6 +1,5 @@
-import { computeHmac, ripemd160 } from '@fuel-ts/crypto';
+import { computeHmac, ripemd160, sha256 } from '@fuel-ts/crypto';
 import { ErrorCode, FuelError } from '@fuel-ts/errors';
-import { sha256 } from '@fuel-ts/hasher';
 import { bn, toBytes, toHex } from '@fuel-ts/math';
 import type { BytesLike } from '@fuel-ts/utils';
 import { arrayify, hexlify, concat, dataSlice, encodeBase58, decodeBase58 } from '@fuel-ts/utils';
