+++ packages/account/src/providers/transaction-request/hash-transaction.ts
@@ -1,5 +1,6 @@
 import { ZeroBytes32 } from '@fuel-ts/address/configs';
-import { uint64ToBytesBE, sha256 } from '@fuel-ts/hasher';
+import { sha256 } from '@fuel-ts/crypto';
+import { uint64ToBytesBE } from '@fuel-ts/hasher';
 import { bn } from '@fuel-ts/math';
 import { TransactionType, InputType, OutputType, TransactionCoder } from '@fuel-ts/transactions';
 import { concat } from '@fuel-ts/utils';
