+++ packages/account/src/providers/transaction-request/upgrade-transaction-request.ts
@@ -1,5 +1,5 @@
+import { hash } from '@fuel-ts/crypto';
 import { FuelError } from '@fuel-ts/errors';
-import { hash } from '@fuel-ts/hasher';
 import type { BN } from '@fuel-ts/math';
 import {
   TransactionType,
