+++ packages/crypto/src/shared/index.ts
@@ -1,3 +1,5 @@
 export * from './scrypt';
 export * from './keccak256';
 export * from './ripemd160';
+export * from './sha256';
+export * from './hashMessage';
