+++ packages/transactions/src/receipt.ts
@@ -1,6 +1,6 @@
 import { BigNumberCoder } from '@fuel-ts/abi-coder';
 import type { AssetId } from '@fuel-ts/address';
-import { sha256 } from '@fuel-ts/hasher';
+import { sha256 } from '@fuel-ts/crypto';
 import type { BN } from '@fuel-ts/math';
 import { arrayify, concat } from '@fuel-ts/utils';
